Merged PR 243: set up MUI and add navbar
- Set up Material UI theme
- Add base router
- Add base NavBar
- Responsive and Accessible

NavBar won't be pushed to vc-kit and is only used in one place and so I've added variables, like logo and app name, inside the component since I don't think it'll be reuse anywhere else. But happy to change the variables into props if needed. Or maybe I should rename it to Header instead.

MUI components have an "sx" prop that allows styling customisation. Originally I used this but then pulled the styles out into styled components. However, I've noticed that some default MUI styling overrode the ones specified in style-components. This is why the Divider component uses "sx". I've added both the use of styled components and "sx" here for your feedback. I'd prefer to use "sx" over styled components just because everything is in one spot. We'll just have to make sure to extract chunks into components to ensure it doesn't become unwieldy. This would be an easy change if we prefer "sx" for all components.

**Desktop**

![Screen Shot 2022-11-01 at 1.03.42 pm.png](https://dev.azure.com/gosource-au/62fbbdb3-5d5e-4897-80b2-1efe68267756/_apis/git/repositories/81185851-2eea-4ea9-8daf-1deb9c13a656/pullRequests/243/attachments/Screen%20Shot%202022-11-01%20at%201.03.42%20pm.png)

 **Mobile**

![Screen Shot 2022-11-01 at 4.47.47 pm.png](https://dev.azure.com/gosource-au/62fbbdb3-5d5e-4897-80b2-1efe68267756/_apis/git/repositories/81185851-2eea-4ea9-8daf-1deb9c13a656/pullRequests/243/attachments/Screen%20Shot%202022-11-01%20at%204.47.47%20pm.png)