Merged PR 3590: chore: setup proxy configuration
The app is now configured to enable node-global-proxy.
If the environment variables HTTP_PROXY and NODE_TLS_REJECT_UNAUTHORIZED are set and the NODE_ENV is either "development" or "test" then the node-global-proxy is used. This set up allows those who do not wish to use proxy to start the app as is without any additional work. Testing in CI pipeline should also work without proxy as long as the environment variables are not set.

With this configuration, all tests are now passing again.

Related work items: #9265