Merged PR 4014: ci: S3 bucket rename
@<Phuc Hoa NGO>  reported an issue when using HA deployment, a 502 error was coming back on issue.

In looking through the payload, I noticed this:
`"@context":["https://www.w3.org/2018/credentials/v1","https://e1-dvp-context.s3.ap-southeast-2.amazonaws.com/AANZFTA-CoO.json"`

This is correct as per the Makefile.  However the actual file should be is:
https://context.e1.nonprod-dvp.cp4.homeaffairs.gov.au/AANZFTA-CoO.json

When looking at the cloudwatch log, I noticed this:

```
2023-02-27T05:15:19.397Z	45002588-6aff-4458-bd0d-97de6d70e789	ERROR	Unhandled Promise Rejection
{
    "errorType": "Runtime.UnhandledPromiseRejection",
    "errorMessage": "jsonld.InvalidUrl: URL \"https://e1-dvp-context.s3.ap-southeast-2.amazonaws.com/AANZFTA-CoO.json\" could not be dereferenced: Not Found",
```

We have an underlying issue here in that pulumi isn't updating the bucket on Makefile change.  I will investigate this.

This PR should repoint the context to a real URL.

Related work items: #12867