Merged PR 4291: ci: check disk space
A surprisingly number of pipeline runs are failing due to lack of space on the agent.  I have tried to get this upgraded but no luck so far.

This PR checks that a minimum of 3GB space is available on the agent as the very first step.  If this is not available, then it errors out.  3GB was chosen due to the size of the node modules directories.

For testing purposes:
When I set the minimum required to 30gb, the outcome was as expected:
https://dev.azure.com/bcz-prod/digital-verification-platform/_build/results?buildId=56852&view=results

When the minimum required was set to 3GB, the outcome was as expected:
https://dev.azure.com/bcz-prod/digital-verification-platform/_build/results?buildId=56853&view=results

Related work items: #14381