Merged PR 4290: ci: admin api
This PR successfully creates the admin api module.  Link to latest successful run:
https://dev.azure.com/bcz-prod/digital-verification-platform/_build/results?buildId=56846&view=results

Notice that for admin-api module, we are using the existing api artifact.  This is because the openapi.json requires paths.  Without this, the api gateway errors out.  @<Ashley HARWOOD> had a good idea to put a placeholder path in, which I tried.  The api / verification code was smart enough to see through this and still errored out unfortunately.  So we are back to using the api artifact as the placeholder.

Related work items: #12957